requirements.txt
torch>=2.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.15.0
tqdm>=4.64.0
numpy>=1.21.0 